Balance changes, fixed those changes so they actually work

+++ Assets/Scripts/PlayerController.cs
@@ -5,7 +5,7 @@ public class PlayerController : MonoBehaviour
 
     public float speed = 5;
     public float health = 100;
-    public int maxHealth = 100;
+    public float maxHealth = 100;
     private CharacterController controller;
     private GameManager gameManager;
     // Start is called before the first frame update
